docs(examples): update package.json and move to components folder
