+++ examples/light-dark-mode/package.json
@@ -9,7 +9,7 @@
     "start": "next start"
   },
   "dependencies": {
-    "@carbon/react": "^1.2.0-rc.0",
+    "@carbon/react": "^1.2.0",
     "next": "12.1.4",
     "react": "18.0.0",
     "react-dom": "18.0.0"
