+++ examples/light-dark-mode/src/components/ThemePreference.js
